Witnessing for vec_len and st_loc (#249)
* examples for witness generation

* abstract a standalone witness package

* witness for st_loc

* assign step


* refactor

* add witness for move/copy_loc

* remove unused

---------

+++ vm-circuit/src/chips/execution_chip_v2/utils.rs
@@ -140,7 +140,7 @@ pub(crate) mod from_limbs {
         let mut value = F::ZERO;
         let mut multiplier = F::ONE;
         for limb in limbs.iter() {
-            value += F::from(*limb as u64) * multiplier;
+            value += F::from(*limb) * multiplier;
             multiplier *= F::from(1u64 << LIMB_BITS);
         }
         value
